test: объединение фикстур
