+++ tests/test_parse.py
@@ -1,50 +1,6 @@
 import os
-from xml.etree import ElementTree as ET
-
-import pytest
 
 from kontur.checkgrammar import parse
-
-# region fixture
-
-
-@pytest.fixture
-def button():
-    obj = ET.parse("tests/fixture/button.xml")
-    return obj
-
-
-@pytest.fixture
-def page_no_title():
-    obj = ET.parse("tests/fixture/noTitle.xml")
-    return obj
-
-
-@pytest.fixture
-def picture():
-    obj = ET.parse("tests/fixture/picture.xml")
-    return obj
-
-
-@pytest.fixture
-def empty():
-    obj = ET.parse("tests/fixture/empty.xml")
-    return obj
-
-
-@pytest.fixture
-def ru_content():
-    obj = ET.parse("tests/fixture/ruContent.xml")
-    return obj
-
-
-@pytest.fixture
-def no_ru_content():
-    obj = ET.parse("tests/fixture/noRuContent.xml")
-    return obj
-
-
-# endregion
 
 
 class TestGetRUContent:
